feat: add cli usage
